--- Licht.Core/Graphics/ICommandList.cs
@@ -1,3 +1,0 @@
-﻿namespace Licht.Core.Graphics;
-
-public interface ICommandList {}
