+++ Licht.GraphicsCore/Graphics/ICommandList.cs
@@ -1,0 +1,3 @@
+﻿namespace Licht.GraphicsCore.Graphics;
+
+public interface ICommandList {}
