+++ Licht.GraphicsCore/Graphics/IRenderer.cs
@@ -1,4 +1,4 @@
-﻿namespace Licht.Core.Graphics;
+﻿namespace Licht.GraphicsCore.Graphics;
 
 public interface IRenderer
 {
