+++ Licht.Vulkan/Licht.Vulkan.csproj
@@ -14,7 +14,7 @@
     </ItemGroup>
 
     <ItemGroup>
-      <ProjectReference Include="..\Licht.Core\Licht.Core.csproj" />
+      <ProjectReference Include="..\Licht.GraphicsCore\Licht.GraphicsCore.csproj" />
     </ItemGroup>
 
 </Project>
